Add screenshot capture for NL page
